resolve file
