+++ monolith/src/main/java/com/skripsi/monolith/controller/user/CountryController.java
@@ -1,7 +1,7 @@
-package com.skripsi.monolith.controller.country;
-import com.skripsi.monolith.model.country.Country;
+package com.skripsi.monolith.controller.user;
+import com.skripsi.monolith.model.user.Country;
 import com.skripsi.monolith.model.user.User;
-import com.skripsi.monolith.service.country.CountryService;
+import com.skripsi.monolith.service.user.CountryService;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.stereotype.Controller;
