+++ monolith/src/main/java/com/skripsi/monolith/controller/user/RoleController.java
@@ -3,22 +3,21 @@ package com.skripsi.monolith.controller.user;
 import com.skripsi.monolith.model.user.Role;
 import com.skripsi.monolith.model.user.User;
 import com.skripsi.monolith.service.user.RoleService;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.stereotype.Controller;
 
 @Controller
-@Slf4j
 public class RoleController {
 
-    private RoleService roleService;
+  private final RoleService roleService;
 
-    public RoleController(RoleService roleService) {
-        this.roleService = roleService;
-    }
+  public RoleController(RoleService roleService) {
+    this.roleService = roleService;
+  }
 
-    @SchemaMapping
-    private Role role(User user) {
-        return roleService.role(user.getRoleId());
-    }
+  @SchemaMapping
+  public Role role(User user) {
+    return roleService.getRole(user.getRoleId());
+  }
+
 }
