--- monolith/src/main/java/com/skripsi/monolith/model/role/Role.java
@@ -1,27 +1,0 @@
-package com.skripsi.monolith.model.role;
-
-import com.skripsi.monolith.model.BaseEntity;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.math.BigInteger;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "roles")
-public class Role extends BaseEntity {
-
-  @Id
-  private BigInteger id;
-
-  private String name;
-
-}
