+++ monolith/src/main/java/com/skripsi/monolith/model/user/Role.java
@@ -6,7 +6,9 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.math.BigInteger;
 
 @Data
@@ -15,8 +17,11 @@ import java.math.BigInteger;
 @AllArgsConstructor
 @Entity
 @Table(name = "roles")
-public class Role{
-    @Id
-    private BigInteger id;
-    private String name;
+public class Role extends BaseEntity {
+
+  @Id
+  private BigInteger id;
+
+  private String name;
+
 }
