+++ monolith/src/main/java/com/skripsi/monolith/repository/user/RoleRepository.java
@@ -1,6 +1,5 @@
 package com.skripsi.monolith.repository.user;
 
-import com.skripsi.monolith.model.user.Country;
 import com.skripsi.monolith.model.user.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
