+++ monolith/src/main/java/com/skripsi/monolith/service/user/CountryService.java
@@ -1,8 +1,8 @@
-package com.skripsi.monolith.service.country;
+package com.skripsi.monolith.service.user;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.skripsi.monolith.model.country.Country;
-import com.skripsi.monolith.repository.country.CountryRepository;
+import com.skripsi.monolith.model.user.Country;
+import com.skripsi.monolith.repository.user.CountryRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
